+++ configs/blinky-analyzer.config
@@ -10,10 +10,12 @@
 
 #-javaagent:$TACOCO_HOME$/lib/blinky-core-0.0.1-SNAPSHOT-jar-with-dependencies.jar=l,frames,sourcename,calldepth
 
+ 
 #-Dtacoco.listeners=org.spideruci.tacoco.testlisteners.JacocoListener
 #-Dtacoco.listeners=org.spideruci.tacoco.testlisteners.BlinkyListener
 -Dtacoco.listeners=org.spideruci.analysis.tacoco.BlinkyListener
 -Dtacoco.analyzer=org.spideruci.tacoco.analysis.TacocoAnalyzer
+# Dtacoco.testIgnore expects list of comma separated name of test classes to be ignored.
 -Dtacoco.testIgnore
 
 #PIT specific options
