+++ configs/pit-analyzer.config
@@ -1,10 +1,11 @@
 # Use these configurations to test tacoco with the spiderMath project in the resources directory,
-# otherwise, set the paths to the exec and jar files appropriately.
+# otherwise, set the paths to the exec and jar files appropriately. 
 -javaagent:$TACOCO_HOME$/lib/org.jacoco.agent-0.7.4.201502262128-runtime.jar=destfile=$OUTDIR$/$PROJECT_NAME$.exec,dumponexit=false
 -Dtacoco.listeners=org.spideruci.tacoco.testlisteners.JacocoListener
 -Dtacoco.analyzer=org.spideruci.tacoco.analysis.PITAnalyzer
 -Dtacoco.db
--Dtacoco.testIgnore
+# Dtacoco.testIgnore expects list of comma seperated name of test classes to be ignored.
+-Dtacoco.testIgnore 
 
 #PIT specific options
 #-Dtacoco.pit.no_unittest_if_result_exist
