+++ configs/tacoco-analyzer.config
@@ -1,7 +1,9 @@
 # Use these configurations to test tacoco with the spiderMath project in the resources directory,
 # otherwise, set the paths to the exec and jar files appropriately.
+
 -javaagent:$TACOCO_HOME$/lib/org.jacoco.agent-0.7.4.201502262128-runtime.jar=destfile=$OUTDIR$/$PROJECT_NAME$.exec,dumponexit=false
 -Dtacoco.listeners=org.spideruci.tacoco.testlisteners.JacocoListener
 -Dtacoco.analyzer=org.spideruci.tacoco.analysis.TacocoAnalyzer
 -Dtacoco.db
--Dtacoco.testIgnore=org.spideruci.benchmark.spiderMath_TestNG.AdditionTest
+# Dtacoco.testIgnore expects list of comma seperated name of test classes to be ignored.
+-Dtacoco.testIgnore=org.spideruci.benchmark.spiderMath_TestNG.AdditionTest, org.spideruci.benchmark.spiderMath_TestNG.MultiplicationTest
